+++ supernode/services/senseregister/task_test.go
@@ -226,7 +226,7 @@ func TestTaskGenFingerprintsData(t *testing.T) {
 				Sexy:     0.5,
 			},
 
-			ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+			ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 			HashOfCandidateImageFile: "HashOfCandidateImageFile",
 		}
@@ -549,7 +549,7 @@ func TestTaskProbeImage(t *testing.T) {
 				Sexy:     0.5,
 			},
 
-			ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+			ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 			HashOfCandidateImageFile: "HashOfCandidateImageFile",
 		}
@@ -877,7 +877,7 @@ func TestTaskValidateDdFpIds(t *testing.T) {
 						Sexy:     0.5,
 					},
 
-					ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+					ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 					HashOfCandidateImageFile: "HashOfCandidateImageFile",
 				},
