+++ walletnode/services/test/sense_service_test.go
@@ -73,7 +73,7 @@ func TestNodesSendImage(t *testing.T) {
 			Sexy:     0.5,
 		},
 
-		ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+		ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 		HashOfCandidateImageFile: "HashOfCandidateImageFile",
 	}
